Fix relative mode editor bindings
